requirements.txt

Flask==2.0.2
Flask_SQLAlchemy
Flask-Bcrypt==0.7.1
SQLAlchemy==1.4.32
bcrypt==3.2.0
PyJWT==2.3.0
psycopg2.binary
sqlalchemy-utils==0.37.7
requests
Werkzeug==2.0.3
Faker
google-cloud-logging